refactor: create new package and reorganize dto
