+++ server/src/site/doget/dto/raw/FnnrRawDto.java
@@ -1,4 +1,4 @@
-package site.doget.dto;
+package site.doget.dto.raw;
 
 import lombok.Getter;
 import lombok.Setter;
