+++ server/src/site/doget/mybatis/mapper/DepositCustomerMapper.xml
@@ -2,7 +2,7 @@
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 
 <mapper namespace="site.doget.mybatis.mapper.DepositCustomerMapper">
-    <resultMap id="DepositCustomerResultMap" type="site.doget.dto.CustomerCountByTypeRawDto">
+    <resultMap id="DepositCustomerResultMap" type="site.doget.dto.raw.CustomerCountByTypeRawDto">
         <result property="newDt" column="new_dt"/>
         <result property="custDscdNm" column="cust_dscd_nm"/>
         <result property="count" column="count"/>
