+++ server/src/site/doget/mybatis/mapper/FinancialMapper.java
@@ -2,10 +2,9 @@ package site.doget.mybatis.mapper;
 
 import org.apache.ibatis.annotations.Mapper;
 import site.doget.dto.FinancialReqDto;
-import site.doget.dto.FnnrRawDto;
+import site.doget.dto.raw.FnnrRawDto;
 
 import java.util.List;
-import java.util.Map;
 
 @Mapper
 public interface FinancialMapper {
