+++ server/src/site/doget/mybatis/mapper/FinancialMapper.xml
@@ -3,7 +3,7 @@
         "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 
 <mapper namespace="site.doget.mybatis.mapper.FinancialMapper">
-    <resultMap id="FinancialResultMap" type="site.doget.dto.FnnrRawDto">
+    <resultMap id="FinancialResultMap" type="site.doget.dto.raw.FnnrRawDto">
         <result property="baseYm" column="base_ym"/>
         <result property="C10000000" column="C10000000"/>
         <result property="C11000000" column="C11000000"/>
