+++ server/src/site/doget/mybatis/mapper/IncomeMapper.java
@@ -2,7 +2,7 @@ package site.doget.mybatis.mapper;
 
 import org.apache.ibatis.annotations.Mapper;
 import site.doget.dto.BankReqDto;
-import site.doget.dto.PlosdocRawDto;
+import site.doget.dto.raw.PlosdocRawDto;
 
 import java.util.List;
