+++ server/src/site/doget/mybatis/mapper/IncomeMapper.xml
@@ -2,7 +2,7 @@
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 
 <mapper namespace="site.doget.mybatis.mapper.IncomeMapper">
-    <resultMap id="IncomeResultMap" type="site.doget.dto.PlosdocRawDto">
+    <resultMap id="IncomeResultMap" type="site.doget.dto.raw.PlosdocRawDto">
         <result property="baseYm" column="base_ym"/>
         <result property="C40000000" column="C40000000"/>
         <result property="C41000000" column="C41000000"/>
