+++ server/src/site/doget/mybatis/mapper/LoanCustomerMapper.xml
@@ -2,12 +2,12 @@
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 
 <mapper namespace="site.doget.mybatis.mapper.LoanCustomerMapper">
-    <resultMap id="CustTypeResultMap" type="site.doget.dto.CustomerCountByTypeRawDto">
+    <resultMap id="CustTypeResultMap" type="site.doget.dto.raw.CustomerCountByTypeRawDto">
         <result property="newDt" column="new_dt"/>
         <result property="custDscdNm" column="cust_dscd_nm"/>
         <result property="count" column="count"/>
     </resultMap>
-    <resultMap id="CustomerCountByAmountResultMap" type="site.doget.dto.CustomerCountByAmountRawDto">
+    <resultMap id="CustomerCountByAmountResultMap" type="site.doget.dto.raw.CustomerCountByAmountRawDto">
         <result property="range" column="range"/>
         <result property="custDscdNm" column="cust_dscd_nm"/>
         <result property="count" column="count"/>
