+++ server/src/site/doget/service/DepositCustomerService.java
@@ -3,9 +3,9 @@ package site.doget.service;
 import org.apache.ibatis.session.SqlSession;
 import org.apache.ibatis.session.SqlSessionFactory;
 import site.doget.dto.*;
+import site.doget.dto.raw.CustomerCountByTypeRawDto;
 import site.doget.mybatis.SqlSessionFactoryProvider;
 import site.doget.mybatis.mapper.DepositCustomerMapper;
-import site.doget.mybatis.mapper.LoanCustomerMapper;
 
 import java.util.ArrayList;
 import java.util.List;
